[AutoDiff] ensure diff witness original function is alive (#30399)
This situation exposed a bug where DeadFunctionElimination eliminated the
SILFunction for `identity<T>` even though a differentiability witness for it
still existed. This causes deserialization of this module to crash when
trying to deserialize the differentiability witness because it can't find
the original function `identity<T>`.

I fixed it by keeping the original function of differentiability witnesses always alive.

A better solution might be to teach DeadFunctionElimination how to eliminate
differentiability witnesses, because I think that the differentiability witness involved
in this situation is actually dead and could be eliminated, but that would be more
work and I just wanted to fix the immediate problem for now.

Fixes TF-1202 and TF-1104.